Add 3XDH functions and reorganize cryptography includes
